chat design and server ok

+++ scripts/run/ai.sh
@@ -53,6 +53,18 @@ if [ "$1" == "start" ]; then
     exit 0
 fi
 
+if [ "$1" == "start-client" ]; then
+	echo "Node JS client is available at http://localhost:5111"
+	docker exec -it -u 1000 fs-ai-node npm run dev:client
+    exit 0
+fi
+
+if [ "$1" == "start-server" ]; then
+	echo "Node JS client is available at http://localhost:5112"
+	docker exec -it -u 1000 fs-ai-node npm run dev:server
+    exit 0
+fi
+
 if [ "$1" == "stop" ]; then
 	docker exec -it fs-ai-node pkill node || true
 	docker exec -it fs-ai-node pkill npm || true
